+++ app/static/img/background.jpg
