+++ app/static/img/wave.png
