+++ app/static/js/theme.js
