+++ run.py
@@ -1,4 +1,5 @@
-from routes import app
+from app import app
+
 
 if __name__ == "__main__":
     app.run(debug=True)
